// types/extended.ts
export interface CalendarItem {
  title: string;
  taskDate: {
    id: string;
    from: Date | undefined;
    to: Date | undefined;
  } | null;
  workspaceId: string;
  workspaceName: string;
  workspaceColor: CustomColors;
  taskId: string;
}